updating Checkout.tsx instead of index.tsx for better file update module path accordingly
